requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.40
psycopg2-binary==2.9.11
python-dotenv
Pillow
pillow-heif
gTTS
requests
soundfile
librosa
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
torchaudio --index-url https://download.pytorch.org/whl/cpu
numpy==1.26.4
scipy
transformers
lhotse
vocos
tensorboard